+++ src/turbulence.rs
@@ -6,10 +6,10 @@ use schemars::{JsonSchema};
 #[serde(rename_all = "camelCase")]
 #[serde(rename = "TurbulencePointJSON")]
 pub struct TurbulencePoint {
-    pub diameter: f32,
-    pub position: f32,
-    pub start_time: f32,
-    pub end_time: f32,
+    pub diameter: f64,
+    pub position: f64,
+    pub start_time: f64,
+    pub end_time: f64,
 }
 
 impl Default for TurbulencePoint {
@@ -18,7 +18,7 @@ impl Default for TurbulencePoint {
             diameter: 0.0,
             position: 0.0,
             start_time: 0.0,
-            end_time: f32::NAN,
+            end_time: f64::NAN,
         }
     }
 }
